requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
et-xmlfile==1.1.0
idna==3.3
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
XlsxWriter==3.0.3

thefuzz~=0.19.0
python-Levenshtein~=0.12.2
